Moved the examples out of their nested include folder to the main include
