+++ src/examples/breath/src/breath.cc
@@ -1,4 +1,4 @@
-#include <breath/breath.h>
+#include "breath.h"
 
 namespace smov {
