+++ src/examples/set_legs_distance_to/src/set_legs_distance_to.cc
@@ -1,4 +1,4 @@
-#include <set_legs_distance_to/set_legs_distance_to.h>
+#include "set_legs_distance_to.h"
 
 namespace smov {
